move and rename navigation bar widget.
