+++ keysmith/lib/src/features/navigation/view/components/navigation_bar_widget.dart
